Add files via upload
This is work-in-progress policy to validate uncompressed quicktime files created in vrecord. It will need to be integrated with the work-in-progress vrecord_policy.xml, (there will be some redundancies) and then additional rules will need to be added to account for other types of files created in vrecord. This work will be picked up by another CUNY intern in 2018.