png of QCTools data
The script uses gnuplot to produce a png of the QCTools data as a means
of quickly assessing a recording. It is still in progress. There are
some issues setting the range of with the y-axis and palette gradient
in a way that is meaningful and will change to reflect the recording’s
specifications, the x-axis is just in seconds, and the position of the
plot’s keys are not quite right.