Updates to png QCTools data
In addition to the suggestions made the image is now a jpeg to
accommodate the need for multiple color ranges. A plot of the audio
phase  data has been added. The yaxis scales are set for the audio
levels and broadcast range. The yaxis and the color gradient for the
saturation will scale to match the video’s bit depth. The title changes
to match the recording’s identifier.