Updates to QC graph Version 2 and documentation: (#373)
* Updates to QC graph Version 2:

Updates in version 2 include:
-The version number is now included into the Graph’s title
-The color range for the plot of %Outside of Broadcast Range was
widened from 0-3% to 0-5%. The new range defines 0 as green, 2% as
yellow, and 5% as red
-The label of for each plot has been moved to the top center of each
plot and displayed over the plotted data
-Minor changes have been made to fonts, size, and position of scales
and keys
Additions were also made to the documentation page for Analog
Digitization, which now includes a description of the QC Graphs
function and an example of the output. The documentation for
Installation and Setup now also includes a note about installing
gnuplot to enable the QC graph functionality.