Capture configuration gui
This is still very much in progresses but it introduces a capture
configuration gui that can be accessed via the edit gui. Configuration
profiles are created, then saved locally, and then the user returns to
the edit gui. Ideally, a dropdown menu in the edit gui would also allow
you to select preexisting profiles, but this has not been added yet.
The configuration profile is then appended to the config_file.